% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey_tables_and_ci.R
\name{create_survey_tables_and_ci}
\alias{create_survey_tables_and_ci}
\title{Generate Survey Summary Statistics}
\usage{
create_survey_tables_and_ci(
  continuous_vars,
  categorical_vars,
  ci_vars,
  survey_design
)
}
\arguments{
\item{continuous_vars}{A character vector of continuous variable names to include in the summary statistics.}

\item{categorical_vars}{A character vector of categorical variable names to include in the summary statistics.}

\item{ci_vars}{A character vector of variable names for which 95\% confidence intervals should be calculated.}

\item{survey_design}{A survey design object created using the \code{survey} package.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{continuous_summary}: A \code{svyCreateTableOne} object summarizing the continuous variables.
\item \code{categorical_summary}: A \code{svyCreateCatTable} object summarizing the categorical variables.
\item \code{confidence_intervals}: A data frame with 95\% confidence intervals for the specified variables, including estimates, lower bounds, and upper bounds.
}
}
\description{
Adding to the \code{survey} package, this function generates summary statistics
for continuous and categorical variables and computes 95\% confidence intervals for specified variables.
The function supports both NHANES and NAMCS datasets with specific
variables for weights, strata, and primary sampling unit (PSU).
}
\examples{
# Example 1: Pre-pandemic 2017-2020 NHANES Data
nhanes_path <- "~/Desktop"
nhanes_files <- c("P_DEMO.XPT","P_ALQ.XPT", "P_BIOPRO.XPT", "P_BMX.XPT", "P_BPQ.XPT", "P_CBC.XPT")
nhanes_data <- load_nhanes_data(nhanes_path, nhanes_files)

# Define NHANES survey design
nhanes_survey_design <- svydesign(
  data = nhanes_data,
  id = ~SDMVPSU,  # Primary sampling unit
  strata = ~SDMVSTRA,  # Stratum
  weights = ~WTINTPRP,  # Survey weights
  nest = TRUE
)

# List out variables for NHANES - age, sex, BMI, and race variables
continuous_vars_nhanes <- c("RIDAGEYR", "BMXBMI")
categorical_vars_nhanes <- c("RIAGENDR", "RIDRETH1")
ci_vars_nhanes <- c("RIAGENDR", "RIDRETH1")

# Generate results for NHANES
nhanes_results <- create_survey_tables_and_ci(
  continuous_vars = continuous_vars_nhanes,
  categorical_vars = categorical_vars_nhanes,
  ci_vars = ci_vars_nhanes,
  survey_design = nhanes_survey_design
)

# Example 2: 2018 NAMCS Data
namcs_path <- "~/Desktop"
namcs_files <- c("namcs2018-stata.dta", "namcs2016-stata.dta")
namcs_data <- load_namcs_data(namcs_path, namcs_files)

# Define NAMCS survey design
namcs_survey_design <- svydesign(
  data = namcs_data,
  id = ~PSU,  # Primary sampling unit
  strata = ~STRATUM,  # Stratum
  weights = ~WEIGHT,  # Survey weights
  nest = TRUE
)

# List out variables for NAMCS - age, sex, BMI, and race variables
continuous_vars_namcs <- c("AGE", "BMI")
categorical_vars_namcs <- c("SEX", "RACERETH")
ci_vars_namcs <- c("SEX", "RACERETH")

# Generate results for NAMCS
namcs_results <- create_survey_tables_and_ci(
  continuous_vars = continuous_vars_namcs,
  categorical_vars = categorical_vars_namcs,
  ci_vars = ci_vars_namcs,
  survey_design = namcs_survey_design
)

}
