% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_namcs_data.R
\name{load_namcs_data}
\alias{load_namcs_data}
\title{Download, Unzip, and Read NAMCS Stata Data}
\usage{
load_namcs_data(year, dest_dir = tempdir(), write_csv = FALSE)
}
\arguments{
\item{year}{The year of the NAMCS dataset to download (e.g., 2009, 2018).}

\item{dest_dir}{The directory to save the downloaded zip file and unzip it (default is a temporary directory).}

\item{write_csv}{A logical value indicating whether to save the data as a CSV file (default is \code{FALSE}).}
}
\value{
A dataframe containing the NAMCS data for the specified year, read from the Stata file.
}
\description{
This function downloads the NAMCS Stata dataset for a specified year from the CDC FTP server,
unzips the file, and loads the data into an R dataframe. Optionally, the function can save the
data as a CSV file for further cleaning.
Zip files can be viewed here: \href{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/dataset_documentation/namcs/stata/}{NAMCS Stata Datasets}
}
\details{
The function constructs a URL to download the NAMCS Stata dataset for the specified year,
unzips the downloaded file, and reads the Stata file into an R dataframe. If \code{write_csv} is \code{TRUE},
the data is saved as a CSV file in the specified directory.
}
\examples{
# Download and load the NAMCS 2018 dataset without writing to CSV
namcs2018 <- load_namcs_data(2018, write_csv = FALSE)

# Download and load the NAMCS 2020 dataset and save as CSV
namcs2020 <- load_namcs_data(2009, write_csv = TRUE)

}
