% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nhanes_data.R
\name{load_nhanes_data}
\alias{load_nhanes_data}
\title{Load and Merge Specific NHANES Datasets}
\usage{
load_nhanes_data(path, files)
}
\arguments{
\item{path}{A string specifying the directory containing the NHANES datasets.}

\item{files}{A character vector of file names to load and merge.}
}
\value{
A data frame containing the merged datasets.
}
\description{
This function loads NHANES datasets from specified files and merges them by \code{SEQN}.
For downloading interested NHANES datasets, visit the official \href{https://wwwn.cdc.gov/nchs/nhanes/}{CDC NHANES website}.
}
\examples{
# Mac-specific example (assuming files are on the Desktop)
nhanes_path_mac <- "~/Desktop"
nhanes_files <- c("P_ALQ.XPT", "P_BIOPRO.XPT", "P_BMX.XPT", "P_BPQ.XPT", "P_CBC.XPT", "P_DEMO.XPT")
nhanes_data_mac <- load_nhanes_data(nhanes_path_mac, nhanes_files)

# Non-Mac (Windows/Linux) specific example
nhanes_path_non_mac <- "C:/Users/YourName/Documents/nhanes_data"
nhanes_files_non_mac <- c("P_ALQ.XPT", "P_BIOPRO.XPT", "P_BMX.XPT", "P_BPQ.XPT", "P_CBC.XPT")
nhanes_data_non_mac <- load_nhanes_data(nhanes_path_non_mac, nhanes_files_non_mac)

}
